feat: restructure user module by organizing controllers and use cases, and update error handling
