+++ src/modules/users/__tests__/CreateUserUseCase.spec.ts
@@ -1,6 +1,6 @@
-import { AppError } from '../../../../errors/AppError'
-import { InMemoryUserRepository } from '../../repositories/inMemory/InMemoryUserRepository'
-import { CreateUserUseCase } from './CreateUserUseCase'
+import { AppError } from '../../../shared/errors/AppError'
+import { InMemoryUserRepository } from '../repositories/inMemory/InMemoryUserRepository'
+import { CreateUserUseCase } from '../useCases/createUser/CreateUserUseCase'
 
 let createUserUseCase: CreateUserUseCase
 let inMemoryUserRepository: InMemoryUserRepository
