+++ src/modules/users/controllers/RemoveUserController.ts
