+++ src/modules/users/controllers/UpdateUserAvatarController.ts
