+++ src/modules/users/useCases/createUser/CreateUserUseCase.ts
@@ -1,5 +1,5 @@
 import { inject, injectable } from 'tsyringe'
-import { AppError } from '../../../../errors/AppError'
+import { AppError } from '../../../../shared/errors/AppError'
 import { User } from '../../entities/User'
 import { IUsersRepository } from '../../repositories/IUsersRepository'
 import { ICreateUserDTO } from '../../dtos'
@@ -8,8 +8,8 @@ import { ICreateUserDTO } from '../../dtos'
 class CreateUserUseCase {
   constructor(
     @inject('UsersRepository')
-    private usersRepository: IUsersRepository
-  ) { }
+    private usersRepository: IUsersRepository,
+  ) {}
 
   async execute({ name, email, company, password, avatar }: ICreateUserDTO): Promise<User> {
     const userExists = await this.usersRepository.findUserByEmail(email)
