+++ src/modules/users/useCases/updateUser/UpdateUserUseCase.ts
@@ -1,5 +1,5 @@
 import { inject, injectable } from 'tsyringe'
-import { AppError } from '../../../../errors/AppError'
+import { AppError } from '../../../../shared/errors/AppError'
 import { User } from '../../entities/User'
 import { IUsersRepository } from '../../repositories/IUsersRepository'
 import { IUpdateUserDTO } from '../../dtos'
