+++ src/modules/users/useCases/updateUserAvatar/UpdateUserAvatarUseCase.ts
@@ -1,6 +1,6 @@
 import { inject, injectable } from 'tsyringe'
+import { AppError } from '../../../../shared/errors/AppError'
 import { IUsersRepository } from '../../repositories/IUsersRepository'
-import { AppError } from '../../../../errors/AppError'
 import { deleteFile } from '../../../../utils/file'
 
 interface IRequest {
