+++ src/routes/users.routes.ts
@@ -1,15 +1,15 @@
 import { Router } from 'express'
 import multer from 'multer'
 
-import { CreateUserController } from '../modules/users/useCases/createUser/CreateUserController'
-import { ListUsersController } from '../modules/users/useCases/listUsers/ListUsersController'
-import { RemoveUserController } from '../modules/users/useCases/removeUser/RemoveUserController'
-import { UpdateUserController } from '../modules/users/useCases/updateUser/UpdateUserController'
-import { AuthenticateUserController } from '../modules/users/useCases/authenticateUser/AuthenticateUserController'
+import { CreateUserController } from '../modules/users/controllers/CreateUserController'
+import { ListUsersController } from '../modules/users/controllers/ListUsersController'
+import { RemoveUserController } from '../modules/users/controllers/RemoveUserController'
+import { UpdateUserController } from '../modules/users/controllers/UpdateUserController'
+import { AuthenticateUserController } from '../modules/users/controllers/AuthenticateUserController'
 
-import { ensureAuthenticated } from '../middlewares/ensureAuthenticated'
+import { ensureAuthenticated } from '../shared/middlewares/ensureAuthenticated'
 
-import { UpdateUserAvatarController } from '../modules/users/useCases/updateUserAvatar/UpdateUserAvatarController'
+import { UpdateUserAvatarController } from '../modules/users/controllers/UpdateUserAvatarController'
 import uploadConfig from '../config/upoad'
 
 const usersRoutes = Router()
@@ -34,7 +34,7 @@ usersRoutes.patch(
   '/:id/avatar',
   ensureAuthenticated,
   uploadAvatar.single('avatar'),
-  updateUserAvatarController.handle
+  updateUserAvatarController.handle,
 )
 
 export { usersRoutes }
